+++ source/packages/@aws-accelerator/accelerator/lib/stacks/network-stacks/network-prep-stack/tgw-resources.ts
@@ -25,7 +25,7 @@ import { pascalCase } from 'pascal-case';
 import { AcceleratorStackProps } from '../../accelerator-stack';
 import { LogLevel } from '../network-stack';
 import { getTgwConfig, getTgwRouteTableId, getTransitGatewayId } from '../utils/getter-utils';
-import { isIpv4 } from '../utils/validation-utils';
+import { isIpAddress } from '../utils/validation-utils';
 import { NetworkPrepStack } from './network-prep-stack';
 
 export class TgwResources {
@@ -256,7 +256,7 @@ export class TgwResources {
       ? customerGateways.filter(cgw => cgw.vpnConnections?.filter(vpn => tgwNames.includes(vpn.transitGateway ?? '')))
       : [];
     const crossAcctFirewallReferenceCgws = tgwVpnCustomerGateways.filter(
-      cgw => !isIpv4(cgw.ipAddress) && !this.stack.firewallVpcInScope(cgw),
+      cgw => !isIpAddress(cgw.ipAddress) && !this.stack.firewallVpcInScope(cgw),
     );
 
     for (const crossAcctCgw of crossAcctFirewallReferenceCgws) {